Split dev server code from build code
Make the build file a little more manageable.
