+++ lib/build.js
@@ -3,21 +3,15 @@
 const path = require('path')
 
 // npm dependencies
-const chokidar = require('chokidar')
-const colour = require('ansi-colors')
 const del = require('del')
 const fse = require('fs-extra')
-const nodemon = require('nodemon')
 const sass = require('sass')
 
 // local dependencies
 const plugins = require('./plugins/plugins')
-const utils = require('./utils')
 const {
   projectDir,
-  packageDir,
 
-  appDir,
   appSassDir,
 
   libSassDir,
@@ -49,16 +43,7 @@ function generateAssetsSync ({ verbose } = {}) {
   proxyUserSassIfItExists('application.scss')
   proxyUserSassIfItExists('settings.scss')
 
-  generateCss(appSassDir, publicCssDir, appSassOptions)
-  generateCss(libSassDir, publicCssDir, libSassOptions)
-}
-
-async function buildWatchAndServe () {
-  generateAssetsSync()
-  await utils.waitUntilFileExists(path.join(publicCssDir, 'application.css'), 5000)
-  watchBeforeStarting()
-  const nodemon = await runNodemon()
-  watchAfterStarting(nodemon)
+  generateCssSync()
 }
 
 function clean () {
@@ -102,7 +87,7 @@ function proxyUserSassIfItExists (filename) {
   fse.writeFileSync(path.join(proxyFilePath), proxyFileLines.join('\n'))
 }
 
-function generateCss (sassPath, cssPath, options = {}) {
+function _generateCssSync (sassPath, cssPath, options = {}) {
   const { filesToSkip = [], filesToRename = {} } = options
   if (!fse.existsSync(sassPath)) return
   fse.mkdirSync(cssPath, { recursive: true })
@@ -126,63 +111,9 @@ function generateCss (sassPath, cssPath, options = {}) {
     })
 }
 
-function watchSass (sassPath, generateSassPath, cssPath, options) {
-  if (!fse.existsSync(sassPath)) return
-  chokidar.watch(sassPath, {
-    ignoreInitial: true,
-    disableGlobbing: true // Prevents square brackets from being mistaken for globbing characters
-  }).on('all', () => {
-    generateCss(generateSassPath, cssPath, options)
-  })
-}
-
-function watchBeforeStarting () {
-  watchSass(appSassDir, libSassDir, publicCssDir, libSassOptions)
-  watchSass(appSassDir, appSassDir, publicCssDir, appSassOptions)
-}
-
-function watchAfterStarting (nodemon) {
-  plugins.watchPlugins(({ missing, added }) => {
-    generateAssetsSync()
-    if (missing.length) {
-      console.log(`Removed ${missing.join(', ')}`)
-    }
-    if (added.length) {
-      console.log(`Added ${added.join(', ')}`)
-    }
-    console.log('Attempting to restart nodemon')
-    nodemon.emit('restart')
-  })
-}
-
-function runNodemon () {
-  // Warn about npm install on crash
-  const onCrash = () => {
-    console.log(colour.cyan('[nodemon] For missing modules try running `npm install`'))
-  }
-
-  // Remove port.tmp if it exists
-  const onQuit = () => {
-    try {
-      fse.unlinkSync(portTmpFile)
-    } catch (e) {}
-    console.log('quit')
-    process.exit(0)
-  }
-
-  return nodemon({
-    watch: [
-      path.join(projectDir, '.env'),
-      path.join(appDir, '**', '*.js'),
-      path.join(appDir, '**', '*.json')
-    ],
-    script: path.join(packageDir, 'listen-on-port.js'),
-    ignore: [
-      'app/assets/*'
-    ]
-  })
-    .on('crash', onCrash)
-    .on('quit', onQuit)
+function generateCssSync () {
+  _generateCssSync(appSassDir, publicCssDir, appSassOptions)
+  _generateCssSync(libSassDir, publicCssDir, libSassOptions)
 }
 
 function autoImportComponentsSync () {
@@ -208,6 +139,6 @@ function autoImportComponentsSync () {
 }
 
 module.exports = {
-  buildWatchAndServe,
-  generateAssetsSync
+  generateAssetsSync,
+  generateCssSync
 }
